#528 refactor makefile; refactor tests; refactor enclave.edl
