+++ TestKeyGenerator.h
@@ -1,0 +1,34 @@
+/*
+    Copyright (C) 2019-Present SKALE Labs
+
+    This file is part of sgxwallet.
+
+    sgxwallet is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as published
+    by the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    sgxwallet is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with sgxwallet. If not, see <https://www.gnu.org/licenses/>.
+*/
+
+#pragma once
+
+#include "stubclient.h"
+
+#include <string>
+#include <vector>
+
+namespace TestKeyGenerator {
+
+void generateDkgKeys(StubClient &client, int n, int t,
+                     std::vector<std::string> &ecdsaKeyNames,
+                     std::vector<std::string> &blsKeyNames, int schainID,
+                     int dkgID);
+
+} // namespace TestKeyGenerator
