+++ configure.ac
@@ -18,14 +18,15 @@ AC_PROG_RANLIB
 
 SGX_INIT()
 
+# Register a command-line option to enable test-only ECALL interfaces in the enclave.
+# '--enable-sgx-test-ecalls'
 AC_ARG_ENABLE([sgx-test-ecalls],
 	[AS_HELP_STRING([--enable-sgx-test-ecalls],
 		[Build test-only enclave ECALL interface])],
 	[enable_sgx_test_ecalls=$enableval],
 	[enable_sgx_test_ecalls=no])
-
 AM_CONDITIONAL([SGX_ENABLE_TEST_ECALLS],
-	[test "x$enable_sgx_test_ecalls" = "xyes"])
+	[test "$enable_sgx_test_ecalls" = "yes"])
 
 gmpdir=${PWD}/gmp-build
 tgmpdir=${PWD}/tgmp-build
