+++ secure_enclave/secure_enclave_common.edl
@@ -179,14 +179,6 @@ enclave {
                                 [out, count = 3072] uint8_t* encrypted_dkg_secret,
                                 [out] uint64_t * enc_len, size_t _t);
 
-        public void trustedDecryptDkgSecret (
-                                [out] int *errStatus,
-                                [out, count = SMALL_BUF_SIZE] char* err_string,
-                                [in, count = 3050] uint8_t* encrypted_dkg_secret,
-                                uint64_t enc_len,
-                                [out, count = 3072] uint8_t* decrypted_dkg_secret
-                                );
-
         public void trustedGetEncryptedSecretShare(
                                 [out]int *errStatus,
                                 [out, count = SMALL_BUF_SIZE] char *err_string,
